Refactor Tests (Split Out Staging and Production Regression Tests).
