+++ tests/mainnet-production-metadata.regression.test.js
@@ -181,166 +181,6 @@ const cadenceTestingSuite =
 
         describe("Metadata testing", () => {
             test.each(totalGetNFTIDsTestCases)(
-                "Mainnet: getNFTIDs - Should be consistent between v1 and staging - %s",
-                async (name) => {
-                    const address = getNFTIDsContractToAddresses[name];
-                    if (address === undefined) {
-                        // Associated test was not implemented.  "Pass" the test for now until we improve coverage.
-                        // console.warn(`${name} lacking test coverage.`);
-                        return;
-                    }
-                    let errorV1,
-                        errorStaging,
-                        scriptResponseV1,
-                        scriptResponseStaging;
-                    try {
-                        scriptResponseV1 = await fcl.decode(
-                            await fcl.send([
-                                fcl.script(getNFTIDsV1),
-                                fcl.args([fcl.arg(address, Address)]),
-                            ])
-                        );
-                    } catch (err) {
-                        errorV1 = err.message
-                            .match(/\[Error Code.*\](.*(\s*error: (.*))?)/m)[0]
-                            .trim();
-                    }
-                    try {
-                        scriptResponseStaging = await fcl.decode(
-                            await fcl.send([
-                                fcl.script(getNFTIDsStaging1),
-                                fcl.args([fcl.arg(address, Address)]),
-                            ])
-                        );
-                        scriptResponseStaging = {
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTIDsStaging2),
-                                    fcl.args([fcl.arg(address, Address)]),
-                                ])
-                            )),
-                        };
-                        scriptResponseStaging = {
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTIDsStaging3),
-                                    fcl.args([fcl.arg(address, Address)]),
-                                ])
-                            )),
-                        };
-                        scriptResponseStaging = {
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTIDsStaging4),
-                                    fcl.args([fcl.arg(address, Address)]),
-                                ])
-                            )),
-                        };
-                    } catch (err) {
-                        errorStaging = err.message
-                            .match(/\[Error Code.*\](.*(\s*error: (.*))?)/m)[0]
-                            .trim();
-                    }
-                    if (errorV1 || errorStaging) {
-                        expect(errorV1).toEqual(errorStaging);
-                    } else {
-                        expect(scriptResponseV1).toEqual(scriptResponseStaging);
-                    }
-                },
-                10000
-            );
-
-            test.each(totalGetNFTsTestCases)(
-                "Mainnet: getNFTs - Should be consistent between v1 and staging - %s",
-                async (name) => {
-                    const args = getNFTsArgs[name];
-                    if (args === undefined) {
-                        // Associated test was not implemented.  "Pass" the test for now until we improve coverage.
-                        // console.warn(`${name} lacking test coverage.`);
-                        return;
-                    }
-                    const fclArgs = fcl.args([
-                        fcl.arg(args[0].value, Address),
-                        fcl.arg(
-                            args[1].value.map((raw) => {
-                                return {
-                                    key: raw.key.value,
-                                    value: raw.value.value.map((rawValue) =>
-                                        parseInt(rawValue.value)
-                                    ),
-                                };
-                            }),
-                            Dictionary({
-                                key: tString,
-                                value: tArray(UInt64),
-                            })
-                        ),
-                    ]);
-                    let errorV1,
-                        errorStaging,
-                        scriptResponseV1,
-                        scriptResponseStaging;
-                    try {
-                        scriptResponseV1 = await fcl.decode(
-                            await fcl.send([fcl.script(getNFTsV1), fclArgs])
-                        );
-                    } catch (err) {
-                        errorV1 = err.message
-                            .match(/\[Error Code.*\](.*(\s*error: (.*))?)/m)[0]
-                            .trim();
-                    }
-                    try {
-                        scriptResponseStaging = await fcl.decode(
-                            await fcl.send([
-                                fcl.script(getNFTsStaging1),
-                                fclArgs,
-                            ])
-                        );
-                        scriptResponseStaging = [
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTsStaging2),
-                                    fclArgs,
-                                ])
-                            )),
-                        ];
-                        scriptResponseStaging = [
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTsStaging3),
-                                    fclArgs,
-                                ])
-                            )),
-                        ];
-                        scriptResponseStaging = [
-                            ...scriptResponseStaging,
-                            ...(await fcl.decode(
-                                await fcl.send([
-                                    fcl.script(getNFTsStaging4),
-                                    fclArgs,
-                                ])
-                            )),
-                        ];
-                    } catch (err) {
-                        errorStaging = err.message
-                            .match(/\[Error Code.*\](.*(\s*error: (.*))?)/m)[0]
-                            .trim();
-                    }
-                    if (errorV1 || errorStaging) {
-                        expect(errorV1).toEqual(errorStaging);
-                    } else {
-                        expect(scriptResponseV1).toEqual(scriptResponseStaging);
-                    }
-                },
-                10000
-            );
-
-            test.each(totalGetNFTIDsTestCases)(
                 "Mainnet: getNFTIDs - Should be consistent between staging and production - %s",
                 async (name) => {
                     const address = getNFTIDsContractToAddresses[name];
